refactor: OrderTables 일급 컬렉션 적용 및 변경
